+++ next.config.js
@@ -1,6 +1,9 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: false,
+  experimental: {
+    appDir: true
+  },
   webpack: (config) => {
     config.module.rules.push({
       test: /\.svg$/,
