+++ src/app/(withheader)/head.tsx
@@ -1,0 +1,9 @@
+export default function Head() {
+  return (
+    <>
+      <title>Diplomacy MOE</title>
+      <meta content='width=device-width, initial-scale=1' name='viewport' />
+      <link rel='icon' href='/favicon.ico' />
+    </>
+  );
+}
