+++ src/components/header/TwitterLoginButton.tsx
@@ -1,11 +1,16 @@
-import Link from 'next/link';
+'use client';
+
+import { useRouter } from 'next/navigation';
 
 export const TwitterLoginButton = () => {
+  const router = useRouter();
+  const onClickHandler = () => {
+    router.push('/login');
+  };
+
   return (
-    <>
-      <button className='btn-primary'>
-        <Link href='/login'>Twitterアカウントでログイン</Link>
-      </button>
-    </>
+    <button className='btn-primary' onClick={onClickHandler}>
+      Twitterアカウントでログイン
+    </button>
   );
 };
