--- src/pages/_document.tsx
@@ -1,24 +1,0 @@
-import Document, { Html, Head, Main, NextScript, DocumentContext } from 'next/document';
-
-class MyDocument extends Document {
-  static async getInitialProps(ctx: DocumentContext) {
-    const initialProps = await Document.getInitialProps(ctx);
-    return { ...initialProps };
-  }
-
-  render() {
-    return (
-      <Html>
-        <Head>
-          <link rel='dns-prefetch' href='//www.google.co.jp' />
-        </Head>
-        <body>
-          <Main />
-          <NextScript />
-        </body>
-      </Html>
-    );
-  }
-}
-
-export default MyDocument;
