--- src/pages/index.tsx
@@ -1,20 +1,0 @@
-import Head from 'next/head';
-import Board from '~/components/board/Board';
-
-const Home = () => {
-  return (
-    <div>
-      <Head>
-        <title>Diplomacy MOE</title>
-        <link rel='icon' href='/favicon.ico' />
-      </Head>
-
-      <main>
-        <h1 className='text-3xl font-bold underline'>Diplomacy MOE3 Reboot</h1>
-        <Board boardData={undefined} />
-      </main>
-    </div>
-  );
-};
-
-export default Home;
